Add command for listing deployed fatmen
